Add italic in FontEntry

+++ src/types.rs
@@ -7,7 +7,7 @@ pub struct FontEntry {
 	pub family: String,
 	pub id: String,
 	pub style: String,
-	pub weight: i64,
-	pub stretch: i64,
+	pub weight: i32,
+	pub stretch: i32,
 	pub italic: bool,
 }
